+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 ### Removed
 ### Changed
+
+- Moved the rendering of `deploy/operator.yaml` to the `operator-sdk new` command instead of `operator-sdk build`
+
 ### Fixed
 ### Deprecated
 ### Security